document-qa/requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
easyocr>=1.7.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
FlagEmbedding>=1.2.0
transformers>=4.38.0,<4.52.0
qdrant-client>=1.7.0
openai>=1.0.0
langsmith>=0.1.0
ragas>=0.2.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
